feat(contracts): stawka kandydata → Numeric(10,2) (import zamówień Erste) (#601)
Migracje 0147 (VeloBank #593, Polkomtel #594) przeszły na Numeric(10,2) stawkę
KLIENTA, ale stawka KANDYDATA została Integer. Import zamówień Erste ma kandydata
ze stawką godzinową z połówką (157.5 PLN/h) — nie da się jej zapisać jako Integer.

Migracja 0148:
- scala dwa równoległe heady 0147 w jeden head (down_revision = krotka),
- przełącza rodzinę stawki kandydata na Numeric(10,2) symetrycznie do strony
  klienta: contracts.rate_candidate, contract_candidate_rates.rate (harmonogram),
  candidate_rate_history.rate (historia).

Modele + schematy stawki kandydata → Decimal/float (konwencja per-plik jak na
main: Decimal w client_order/new_contractor_order, float w contract.py).
contract_amendment.new_rate_* też dziesiętne. Logika marży zwraca Decimal→Numeric.

Konwersja int→numeric bezstratna, in-place. Frontend bez zmian (formatCurrency/
fmtMoney renderują połówki; ścieżka Decimal istnieje od strony klienta).

+++ backend/app/models/contract.py
@@ -121,10 +121,9 @@ class Contract(Base, TimestampMixin):
     start_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
     end_date: Mapped[Optional[date]] = mapped_column(Date)
 
-    # Stawki finansowe
-    rate_candidate: Mapped[Optional[int]] = mapped_column(Integer)
-    # Stawka klienta — Numeric(10,2): część klientów (np. VeloBank) ma stawki
-    # godzinowe z groszami (206.25, 162.50). Kandydat trzymany jako Integer.
+    # Stawki finansowe — Numeric(10,2): stawki godzinowe bywają z groszami/
+    # połówką (np. klient VeloBank 206.25, kandydat Erste 157.5).
+    rate_candidate: Mapped[Optional[Decimal]] = mapped_column(Numeric(10, 2))
     rate_client: Mapped[Optional[Decimal]] = mapped_column(Numeric(10, 2))
     # Stawka z umowy ramowej (MSA) — wartość referencyjna uzgodniona w umowie
     # ramowej z klientem. NIE wchodzi do liczenia marży (to baseline/ceiling).
